refactor: move accessors to own module
